protocol: prefer cycling low score peers
Cycling is meant to opportunistically connecting to a new peer
in case we're connected to a malicious subset of the network.
We now prefer dropping a low score peer, since those will be
either ones which already failed newly at least one test, or
newly connected ones at the same height. In particular, they
can't be a newly connected one that's found another chain, or
they'd be in synchronizing state, not normal state.